# SharedQueueWatcher simplified