--- a/src/Microsoft.Azure.WebJobs.Extensions.Storage/Queues/Listeners/SharedQueueWatcher.cs
+++ b/src/Microsoft.Azure.WebJobs.Extensions.Storage/Queues/Listeners/SharedQueueWatcher.cs
@@ -2,21 +2,23 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Microsoft.Azure.WebJobs.Host.Queues.Listeners
 {
     internal class SharedQueueWatcher : IMessageEnqueuedWatcher
     {
-        private readonly ConcurrentDictionary<string, ConcurrentBag<INotificationCommand>> _registrations =
-            new ConcurrentDictionary<string, ConcurrentBag<INotificationCommand>>();
+        private readonly ConcurrentDictionary<string, INotificationCommand[]> _registrations =
+            new ConcurrentDictionary<string, INotificationCommand[]>();
 
         public void Notify(string enqueuedInQueueName)
         {
-            ConcurrentBag<INotificationCommand> queueRegistrations;
+            INotificationCommand[] queueRegistrations;
 
             if (_registrations.TryGetValue(enqueuedInQueueName, out queueRegistrations))
             {
-                foreach (INotificationCommand registration in queueRegistrations.ToArray())
+                foreach (INotificationCommand registration in queueRegistrations)
                 {
                     registration.Notify();
                 }
@@ -25,9 +27,14 @@ namespace Microsoft.Azure.WebJobs.Host.Queues.Listeners
 
         public void Register(string queueName, INotificationCommand notification)
         {
-            _registrations.AddOrUpdate(queueName,
-                new ConcurrentBag<INotificationCommand>(new INotificationCommand[] { notification }),
-                (i, existing) => { existing.Add(notification); return existing; });
+            _registrations.AddOrUpdate(queueName, new[] { notification },
+                (i, existing) =>
+                {
+                    var updated = new INotificationCommand[existing.Length + 1];
+                    existing.CopyTo(updated, 0);
+                    updated[existing.Length] = notification;
+                    return updated;
+                });
         }
     }
 }