--- a/test/Benchmarks/WebjobsExtensionsStorage.cs
+++ b/test/Benchmarks/WebjobsExtensionsStorage.cs
@@ -2,7 +2,6 @@
 using BenchmarkDotNet.Attributes;
 using Microsoft.Azure.Storage.Queue;
 using Microsoft.Azure.WebJobs.Host.Queues;
-using Microsoft.Azure.WebJobs.Host.Queues.Listeners;
 using Newtonsoft.Json.Linq;
 
 namespace Benchmarks
@@ -10,10 +9,8 @@ namespace Benchmarks
     [MemoryDiagnoser]
     public class WebjobsExtensionsStorage
     {
-        const string WatcherQueueName = "test";
         readonly CloudQueueMessage _minimal;
         readonly CloudQueueMessage _moderate;
-        readonly SharedQueueWatcher _watcher;
 
         public WebjobsExtensionsStorage()
         {
@@ -34,10 +31,6 @@ namespace Benchmarks
             };
 
             _moderate = new CloudQueueMessage(moderate.ToString());
-
-            // assume one listener
-            _watcher = new SharedQueueWatcher();
-            _watcher.Register(WatcherQueueName, new NoopNotifyCommand());
         }
 
         [Benchmark]
@@ -50,17 +43,6 @@ namespace Benchmarks
         public Guid? QueueCausalityManager_GetOwner_Moderate()
         {
             return QueueCausalityManager.GetOwner(_moderate);
-        }
-
-        [Benchmark]
-        public void SharedQueueWatcher_Notify()
-        {
-            _watcher.Notify(WatcherQueueName);
-        }
-
-        class NoopNotifyCommand : INotificationCommand
-        {
-            public void Notify() { }
         }
     }
 }