This PR changes the way `SharedQueueWatcher` is written, by replacing the costly `ConcurrentBag` and the call to `.ToArray` when notifing. The replacement is provided with a simple array `INotificationCommand[]`. The update function for `AddOrUpdate` is changed to operate on the input and return a new array. It ensures that the value, if swapped, in the concurrent dictionary will be properly updated.

This removes the `.ToArray` and heaviness of `ConcurrentBag` from the hot `Notify` path. 

### Benchmarks

After introducing this change I understood that it's highly likely that there will be a single notification command and run a bechmark that is in history of these changed, but is removed as it's something that won't be useful in the future.

#### Before:

|                    Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| SharedQueueWatcher_Notify | 150.8 ns | 2.78 ns | 5.15 ns | 0.0100 |     - |     - |      32 B |

#### After:

|                    Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|
| SharedQueueWatcher_Notify | 44.40 ns | 1.757 ns | 4.897 ns |     - |     - |     - |         - |

It's a small gain but makes the code a bit clearer imho, so I'll still propose it as a change.