@brettsam 2020-10-07
Thanks again @Scooletz -- do you have benchmarks to show the improvements on this one? Adding @mathewc as he's been doing a bunch of perf work, specifically in the HttpTrigger pipeline.