@mathewc 2020-10-07
Thanks @Scooletz - these changes look good to me. This is code that hasn't been changed forever, and I see there aren't any unit tests for SharedQueueWatcher. I'd feel safer if along with these changes we had a Mock unit test or two that verified this. E.g. just a test that Registers some notifications, and verifies the correct number are called when Notify is invoked.