@Scooletz 2020-10-08
I rewrote the history and pushed with `-force` to make it look like:

1. add a benchmark to provide some numbers for this PR
1. implementation changed
1. the benchmark removed as it's not vital for future development
1. a concurrent Thread-based test

Looking forward for a review and comments 😄